% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow.R
\name{accumulate_flow}
\alias{accumulate_flow}
\title{Compute river flow based on selected routing method.}
\usage{
accumulate_flow(x, method = "simple")
}
\arguments{
\item{x}{A 'HSrunoff' object obtained by compute_segment_runoff()}

\item{method}{Character string specifying the method to be used.}
}
\value{
Returns the input river network where runoff timeseries is switched to accumulated flow for each river segment.
}
\description{
Compute river flow based on selected routing method.
}
\section{Details}{

Currently only simple addition downstream is implemented, specified by value "simple" in method input.
}

