% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_flow.R
\name{accumulate_flow_instant}
\alias{accumulate_flow_instant}
\title{Compute instantaneous flow for each timestep}
\usage{
accumulate_flow_instant(HSrunoff, verbose = FALSE)
}
\arguments{
\item{HSrunoff}{A 'HSrunoff' object obtained by \code{\link{compute_segment_runoff}}}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns the input river network (class \code{HSflow}) where runoff timeseries is switched to 
the accumulated flow for each river segment.
}
\description{
Applies the simplest possible river routing scheme, instantaenous flow, by adding runoff from each river 
segment to all of the segments downstream, for each timestep.
}
