% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff.R
\name{accumulate_runoff.instant}
\alias{accumulate_runoff.instant}
\title{Compute instantaneous flow for each timestep}
\usage{
\method{accumulate_runoff}{instant}(HSrunoff, verbose = FALSE, ...)
}
\arguments{
\item{HSrunoff}{A 'HSrunoff' object obtained by \code{\link{downscale_runoff}}}

\item{verbose}{Whether or not print progress indicators.}

\item{...}{Arguments passed to the routing algorithm.}
}
\value{
Returns an object of class \code{HSflow}) with 
\itemize{
  \item \code{river} River network inherited from input \code{HSrunoff}.
  \item \code{discharge} A list of tables with estimated discharge for each 
    river segment and each timestep.
 \item \code{Routing_Method} The algorithm used for routing.
}
}
\description{
Applies the simplest possible river routing scheme, instantaenous flow, by 
adding runoff from each river segment to all of the segments downstream, for 
each timestep.
}
