% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff.R
\name{accumulate_runoff.muskingum}
\alias{accumulate_runoff.muskingum}
\title{Implements 'muskingum' routing algorithm for vector river network}
\usage{
\method{accumulate_runoff}{muskingum}(HSrunoff, velocity = 1, x,
  verbose = FALSE, ...)
}
\arguments{
\item{HSrunoff}{A \code{HSrunoff} object.}

\item{velocity}{Flow velocity applied to compute parameter x. Can be a constant,
or a vector of flow values at river segments.}

\item{x}{Value for parameter x.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns an object of class \code{HSflow}) with 
\itemize{
  \item \code{river} River network inherited from input \code{HSrunoff}.
  \item \code{discharge} A list of tables with estimated discharge for each 
    river segment and each timestep.
  \item \code{Routing_Method} The algorithm used for routing.
}
}
\description{
The function implements 'muskingum' routing scheme where the storage parameter
\emph{k} is computed using user input flow velocity, and the length of a 
river segment. Using 'muskingum' for runoff data with time interval between
runoff longer than day may cause instability in the output. If the interval is
too high, use another algorithm.
}
\details{
Warning: The function is experimental and has not been thoroughly tested yet.
}
