% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow.R
\name{accumulate_runoff_addition}
\alias{accumulate_runoff_addition}
\title{Title}
\usage{
accumulate_runoff_addition(x)
}
\arguments{
\item{x}{A 'HSrunoff' object obtained by compute_segment_runoff()}
}
\value{
Returns the input river network where runoff timeseries is switched to accumulated flow for each river segment.
}
\description{
Title
}
