% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff_constant.R
\name{accumulate_runoff_constant}
\alias{accumulate_runoff_constant}
\title{Apply constant velocity river routing}
\usage{
accumulate_runoff_constant(HS, velocity = 1, verbose = FALSE)
}
\arguments{
\item{HS}{A \code{HS} object.}

\item{velocity}{Flow velocity. Can be a constant, or a vector of flow 
velocity at each unique river segments. Defaults to 1 meter per second.}

\item{verbose}{Whether to print progress indication or not.}
}
\value{
Returns the input object \code{HS}) with an added list column
  \code{discharge_ts} containing routed discharge estimates for each river
   segment.
}
\description{
Implements a simple constant velocity routing algorithm which can be used 
with arbitrary timesteps.
}
