% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff_muskingum_cunge.R
\name{accumulate_runoff_muskingum_cunge}
\alias{accumulate_runoff_muskingum_cunge}
\title{Apply Muskingum-Cunge river routing}
\usage{
accumulate_runoff_muskingum_cunge(
  HS,
  manning,
  slope,
  channel_width = "Moody",
  q_ref = NULL,
  celerity = NULL,
  verbose = FALSE
)
}
\arguments{
\item{HS}{A \code{HS} object.}

\item{manning}{Manning's roughness coefficient. Either a constant, a 
vector of values corresponding to each river segment, or the name of 
the column of \code{HS} where the values are stored.}

\item{slope}{River bed slopes. Either a constant, a 
vector of values corresponding to each river segment, or the name of 
the column of \code{HS} where the values are stored.}

\item{channel_width}{Channel widths. See details.}

\item{q_ref}{Reference discharge for which channel width computations are
based on. The default is \code{NULL}, which means that q_ref is automatically
computed from inflow timeseries.}

\item{celerity}{If desired, (constant) wave celerities for each river 
segment can be given. Default \code{NULL} means that wave celerity is
automatically computed for each river segment and each timestep.}

\item{verbose}{Whether or not to print progress information. Defaults to 
\code{FALSE}.}
}
\value{
Returns the input object \code{HS}) with an added list column
  \code{discharge_ts} containing routed discharge estimates for each river
   segment.
}
\description{
The function implements Muskingum routing scheme where the storage parameter
\emph{k} is computed using user input wave celerity, and the length of a 
river segment. Using Muskingum for runoff data with time interval day may 
cause instability in the output. If the interval is too high, it is advised 
to use another routing algorithm.
}
\details{
Muskingum hydrologic routing method consists of two parameters; K and x. K
corresponds to travel time through a river segments, and can be estimated 
using Wave celerity and length of the flow line. x is an empirical parameter
for which the value can be found through trial and error. hydrostreamer 
currently does not offer automated calibration of x.

Channel widths can be given as a vector, however by default channel width
is computed from a power-law relationship from reference discahrge (q_ref). 
If the  \code{length(channel_width) == 2}, the first parameter is alfa and
the second is beta in \code{alfa*q_ref^beta}. "Allen" and "Moody" correspond
to optimised parameters.
}
