% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff.R
\name{accumulate_runoff_simple}
\alias{accumulate_runoff_simple}
\title{Implements a simple lagged river routing algorithm.}
\usage{
accumulate_runoff_simple(HSrunoff, boundary = NULL, velocity = 1,
  verbose = FALSE)
}
\arguments{
\item{HSrunoff}{A \code{HSrunoff} object.}

\item{boundary}{A \code{HSobs} object with boundary condition for routing.}

\item{velocity}{Flow velocity applied to compute parameter x. Can be a constant,
or a vector of flow values at river segments.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns an object of class \code{HSflow}) with 
\itemize{
  \item \code{river} River network inherited from input \code{HSrunoff}.
  \item \code{discharge} A list of tables with estimated discharge for each 
    river segment and each timestep.
  \item \code{Routing_Method} The algorithm used for routing.
}
}
\description{
Implements a simple lag-based routing algorithm which can be used with
arbitrary intervals between runoff units. Includes a possibility to
add boundary condition for river routing at certain river segments.
}
\details{
If a boundary condition is supplied, inflow to the specified river segment 
is set to the value provided in boundary condition. The boundary condition
timeseries must equal the runoff timeseries being routed.
}
