% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_control.R
\name{add_control}
\alias{add_control}
\title{Adds flow controls to an 'HS' object}
\usage{
add_control(HS, timeseries, unit, riverIDs, control, type)
}
\arguments{
\item{HS}{An \code{HS} object.}

\item{timeseries}{a data frame with observations.}

\item{unit}{The unit of values in \code{timeseries}. Should be convertible to
cubic meters per second.}

\item{riverIDs}{A vector of riverID of the river segments of the columns
in obs.}

\item{control}{Type of control, see details.}

\item{type}{Whether the control should apply to \code{runoff} input, or to
\code{discharge} output.}
}
\value{
Returns the \code{HS} object with added list column 
  \code{control_ts} and \code{control_type}.
}
\description{
Adds flow control timeseries to a \code{HS} object. Control 
conditions should cover the entire timeseries in \code{HS}.
}
\details{
Available control schemes are 
\itemize{
  \item set: set input flow at a river segment (override routing). Useful 
    e.g. for boundary conditions or dams.
  \item add: increase runoff at a river segment.
  \item subtract: lower runoff at a river segment.
  \item multiply: modify runoff.
}
}
