% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_HSgrid.R
\name{add_HSgrid}
\alias{add_HSgrid}
\title{Combine \code{HSgrid} objects or addnew runoff input from raster.}
\usage{
add_HSgrid(HSgrid, from = NULL, raster = NULL, date = NULL,
  timestep = NULL, aoi = NULL, name = NULL)
}
\arguments{
\item{HSgrid}{an existing \code{HSgrid} object.}

\item{from}{A \code{HSgrid} object to add from. Optional.}

\item{raster}{A Raster* object.}

\item{date}{Date of the first layer in \code{raster}, or a vector of dates 
with the length of layers in \code{raster}.}

\item{timestep}{Length of timestep in raster. Currently supporting 
\code{"hour"}, \code{"day"}, \code{"month"}. Only needed if \code{date} is
not a vector of dates.}

\item{aoi}{Area of interest as an 'sf' polygon object, 
'SpatialPolygons', or 'SpatialPolygonsDataFrame'. Optional, but 
recommended for all applications.}

\item{name}{A name for the runoff timeseries. If \code{NULL},}
}
\value{
Returns the input \code{HSgrid} object with added runoff.
}
\description{
Combines the runoff lists from two HSgrid objects, or adds a new
runoff table from a raster. \emph{Warning: The geometries of the
combined grids must be identical.}
}
