% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_by_topology.R
\name{arrange_by_topology}
\alias{arrange_by_topology}
\title{Arrange network edges in order of topology upstream to downstream}
\usage{
arrange_by_topology(sf_river_network, riverID = "riverID", verbose = FALSE)
}
\arguments{
\item{sf_river_network}{A LINESTRING simple feature collection.}

\item{riverID}{unique edge identifier.}

\item{verbose}{Boolean. Print messages.}
}
\value{
Returns the input sfc arranged by topology.
}
\description{
A sorting method for arranging tables of sfc LINESTRINGs in topological
order.
}
\examples{
# Miera example
library(dplyr)

 miera = hydrostreamer::miera_rivers \%>\%
     dplyr::mutate("default_order" = dplyr::row_number())

 # miera \%>\%
 # ggplot()+
 #    geom_sf(aes(geometry = geom, col = default_order))

 miera_ordered_by_topology = arrange_by_topology(miera,
                                                   riverID = "riverID",
                                                   verbose = TRUE)
 
 #miera_ordered_by_topology \%>\%
 #    ggplot()+
 #    geom_sf(aes(geometry = geom, col = topo_order))

}
