% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2grid.R
\name{average_monthly_runoff}
\alias{average_monthly_runoff}
\title{Monthly average of a timeseries. Expects the timeseries to start from Januart and end in December.}
\usage{
average_monthly_runoff(x, timesteps = NULL)
}
\arguments{
\item{x}{A 'HSgrid' object obtained with polygrid_timeseries(), or a 'RasterStack', or a 'RasterBrick'.}

\item{timesteps}{A logical vector with same length as there are layers in brick, or a numeric vector
specifying the layers to process. Optional. If not given, will process all raster layers.}
}
\value{
returns either a 'HSgrid' object with monthly summary of the timeseries, or a 'RasterBrick',
  depending on the class of x.
}
\description{
Monthly average of a timeseries. Expects the timeseries to start from Januart and end in December.
}
