% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin}
\alias{basin}
\title{Area of Interest polygon}
\format{An 'sf' POLYGON object with no attributes.
\describe{
  \item{geometry}{'sfc' geometry column}
}}
\usage{
basin
}
\description{
A polygon containing the area of interest, used in the vignette of hydrostreamer
}
\keyword{datasets}
