% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basins}
\alias{basins}
\title{Catchment areas for 41 river segments}
\format{
An 'sf' POLYGON object with no attributes.
\describe{
  \item{SEGMENT_ID}{unique ID}
  \item{geometry}{'sfc' geometry column}
}
}
\usage{
basins
}
\description{
River segment specific catchment areas for 41 river segment in 
"example_rivers" data, also provided in the package.
}
\keyword{datasets}
