% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basins_data.R
\docType{data}
\name{basins_data}
\alias{basins_data}
\title{River basins for river network in Southeast Asia}
\format{
An sf object with 216 rows and 2 variables: \describe{
  \item{SEGMENT_ID}{A unique identifier, numeric}
  \item{geom}{simple feature geometry, sfc_MULTIPOLYGON sfc} ... }
}
\source{
\url{https://portal.opentopography.org/raster?opentopoID=OTALOS.112016.4326.2}
}
\usage{
basins_data
}
\description{
A simple feature dataset containing subbasins for a river network in
Southeast Asia derived from ALOS World 3D at 30 meter resolution DEM (Tanado
et al 2014).
}
\references{
\url{https://doi.org/10.5194/isprsannals-II-4-71-2014}
}
\keyword{datasets}
