% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runoff.R
\name{combine_runoff}
\alias{combine_runoff}
\title{Combines runoff and/or discharge in HS objects}
\usage{
combine_runoff(
  HS,
  weights,
  intercept = 0,
  bias = 0,
  drop = FALSE,
  monthly = FALSE
)
}
\arguments{
\item{HS}{An \code{HS} object}

\item{weights}{A vector of weights with the same length as the number of
runoff timeseries in \code{HS}, or a list of weight vectors.}

\item{intercept}{If weights are associated with an intercept, include it here.
Default intercept = 0.}

\item{bias}{Bias in percentage. Defaults to 0.}

\item{drop}{Whether to remove source timeseries and replace them with 
combinations, or keep source timeseries and add combination as new column.}

\item{monthly}{If provided weights are for each month, set to \code{TRUE}
to apply weights separately for each month.}
}
\value{
Returns the input \code{HS} object where runoff timeseries is 
  replaced by the the combined timeseries'.
}
\description{
The function provides the ability to do a weighted combination of  runoff 
and/or discharge timeseries in \code{HS} object using user provided weights.
Function can process several input weight sets.
}
\details{
If a named list of weights is provided to the function, their names will be 
preserved in the combinations. Otherwise the combinations are named using a 
running number.

If \code{intercept} is provided, it is added to the combined timeseries.
If \code{bias} is provided, the combined timeseries is adjusted by
timeseries / (1-bias/100). 
 
Note that if the timeseries of dates is different in each runoff timeseries
in the input, the output includes only those dates which are present in 
every input timeseries.

The combination is only applied to columns \code{runoff_ts}, and 
\code{discharge_ts}. Columns \code{observation_ts} and 
\code{control_ts} are left untouched if present.
}
\examples{
\dontrun{
  weights <- c(0.2,0.2,0.3,0.3,0)
  combined <- combine_runoff(HS, weights)
  
  listweights <- list(weights1 = c(0.2,0.2,0.3,0.3,0),
                      weights2 = c(0.1,0.2,0.3,0.2,0.2),
                      weights3 = c(0,0,0,0.5,0.5))
 combined2 <- combine_runoff(HS, listweights)
}


}
