% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hydrological_signatures.R
\name{compute_hydrological_signatures}
\alias{compute_hydrological_signatures}
\title{Computes a reference discharge}
\usage{
compute_hydrological_signatures(
  HS,
  fun = mean,
  timeseries = "discharge",
  output_col = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{HS}{a \code{HS} object obtained passed through
\code{\link{river_network}}.}

\item{fun}{Function applied to the collected values upstream. The function
must take \code{na.rm} argument.}

\item{timeseries}{Which timeseries to apply \code{f} to. Eithr "runoff",
"discharge", "observation", or "control".}

\item{output_col}{Column name in \code{HS} where the collected values will
be stored. if NULL, the function returns the collected values as a vector.}

\item{verbose}{Whether or not print progress indicators.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A vector of the collected stats, or the input \code{HS} with a 
  specified column storing the vector.
}
\description{
Function returns the river network consisting of all upstream segments 
from the specified river segment.
}
