% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_network_length.R
\name{compute_network_length}
\alias{compute_network_length}
\title{Computes the maximum upstream network length}
\usage{
compute_network_length(HS, values, output_col = NULL, verbose = FALSE)
}
\arguments{
\item{HS}{a \code{HS} object obtained passed through
\code{\link{river_network}}.}

\item{values}{A vector of values to be aggregated, or a name of the column
in \code{HS} storing the values. If the input is a \code{units} object,
the unit is preserved.}

\item{output_col}{Column name in \code{HS} where the collected values will
be stored. if NULL, the function returns the collected values as a vector.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
A vector of the collected stats, or the input \code{HS} with a 
  specified column storing the vector.
}
\description{
Function returns the maximum length of upstream river network for each 
river segment in the network.
}
