% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_river_weights}
\alias{compute_river_weights}
\title{Assign weights to area features within grid cells.}
\usage{
compute_river_weights(river, grid, seg_weights = "length")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{grid}{A 'HSgrid' object. Required.}

\item{seg_weights}{A character vector specifying type of weights, or a vector of user-specified
weights. See Details section. Defaults to "length".}
}
\value{
Returns an 'sf' linestring feature (which is a union of basins, and grid) with an attribute
  column 'weights'.
}
\description{
Assign weights to area features within grid cells.
}
\section{Details}{

  seg_weights should be one of the following: "equal", "length", "strahler", or a numeric vector
  which specifies the weight for each river segment. "equal" assigned grid cell value to all river
  segments within the cell, equally."length" does the same, but weights are based on the length of
  river segment within the cell, so that longer segments get higher weights. "strahler" computes the
  strahler stream order for the river network and weights are based on that.
}

