% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow.R
\name{compute_runoff_using_area}
\alias{compute_runoff_using_area}
\title{Computes specific runoff generated in the river segment based on area weighting.}
\usage{
compute_runoff_using_area(river, basins, grid, rID = "ID", bID = "rID",
  timesteps = NULL, unit = "mm/s")
}
\arguments{
\item{river}{A 'HSrnet' object obtained with river_network() function.}

\item{basins}{An 'sf' polygon object with basin specific areas. May be obtained with river_voronoi(), delineate_basins(),
or some other means. Must have an ID column corresponding to the river segment IDs, and a weights column.}

\item{grid}{A 'HSgrid' object obtained with polygrid_timeseries() function.}

\item{rID}{Name of the column in river with unique IDs.}

\item{bID}{Name of the column in basins with IDs corresponding to rID.}

\item{timesteps}{A logical vector of the length of timesteps in x, or a numeric vector specifying which timesteps to process. If no timesteps are given, all will be processed.}

\item{unit}{Either "mm/s", or "m3/s", depending on the unit of runoff timeseries in grid.}
}
\value{
Returns input river object (class 'HSrunoff') added with timeseries information as new attribute columns.
}
\description{
Computes specific runoff generated in the river segment based on area weighting.
}
