% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_segment_runoff_without_HSweights.R
\name{compute_segment_runoff_without_HSweights}
\alias{compute_segment_runoff_without_HSweights}
\title{Computes specific runoff generated in the river segment without using a HSweights object.}
\usage{
compute_segment_runoff_without_HSweights(river, weights, grid,
  rID = "riverID", wID = "riverID", timesteps = NULL, unit = "mm/s")
}
\arguments{
\item{river}{River network information.}

\item{weights}{With ID and weights information.}

\item{grid}{A \code{HSgrid} object obtained with \code{\link{polygrid_timeseries}}}

\item{rID}{Name of the column in \code{river} with unique IDs.}

\item{wID}{Name of the column in \code{weights} with IDs corresponding to rID.}

\item{timesteps}{A logical vector of the length of timesteps in x, or a numeric vector specifying which 
timesteps to process. If no timesteps are given, all will be processed.}

\item{unit}{Unit of runoff. Can be either "mm/s", or "m3/s". Defaults to mm/s (equivalent to kg/m2/s).}
}
\value{
Returns input river object (class 'HSrunoff') added with timeseries information as new attribute columns.
}
\description{
This function can be used to compute downscaled, segment specific, unrouted, runoff manually from a selection
of input data. This can be used if for some reason not all information required in previous steps is available.
See details below. Called by \code{\link{compute_segment_runoff}}.
}
\details{
The minimum information needed to compute segment specific runoff are:
\itemize{
  \item{river}{At minimum a data.frame with unique IDs for each river segment}
  \item{weights}{At minimum, a data.frame with columns specifying the river ID corresponding to \code{river}, 
  grid ID corresponding to \code{grid}, and weight used for multiplying runoff value in \code{grid}}
  \item{grid}{A \code{HSgrid} object obtained with \code{\link{polygrid_timeseries}}}
}
In practise using this function is not recommended, and instead the \code{HSweights} object should be constructed
and used.
}
\examples{
\dontrun{
#to be added
} 

}
