% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_upstream_aggregate.R
\name{compute_upstream_aggregate}
\alias{compute_upstream_aggregate}
\title{Computes an aggeregate value for all upstream segments}
\usage{
compute_upstream_aggregate(
  HS,
  values,
  fun = sum,
  output_col = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{HS}{a \code{HS} object obtained passed through
\code{\link{river_network}}.}

\item{values}{A vector of values to be aggregated, or a name of the column
in \code{HS} storing the values. If the input is a \code{units} object,
the unit is preserved.}

\item{fun}{Function applied to the collected values upstream. The function
must take \code{na.rm} argument.}

\item{output_col}{Column name in \code{HS} where the collected values will
be stored. if NULL, the function returns the collected values as a vector.}

\item{verbose}{Whether or not print progress indicators.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A vector of the collected stats, or the input \code{HS} with a 
  specified column storing the vector.
}
\description{
Function returns the river network consisting of all upstream segments 
from the specified river segment.
}
