% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_weights}
\alias{compute_weights}
\title{Computes the weights used for each river segment when assigning values of an arbitrarily shaped polygon
  network. The function computes the weights either based on catchment areas(polygons), or line segments
  which intersect a grid cell. Weights are assigned so that the sum of weights for area or line features
  within a grid cell equals to 1.}
\usage{
compute_weights(river, grid, weights, aoi = NULL, basins = NULL,
  drain.dir = NULL, riverID = "ID")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{grid}{A 'HSgrid' object. Required.}

\item{weights}{A character vector specifying type of weights, or a vector of user-specified weights. See
Details section. Required.}

\item{aoi}{An 'sf' polygon object, or an object which can be converted to 'sf' using st_as_sf(). Optional.}

\item{basins}{If weights are set to "area", providing basins skips the polygon generation process.
See Details section. Optional.}

\item{drain.dir}{If weights are set to "area", providing a drainage direction raster enables delineation
of drainage basins from the provided river network. See Details. Optional.}

\item{riverID}{A character string which specifies the name of the column containing unique identifier in
river network. Defaults to "ID".}
}
\value{
Returns a list object with class 'HSrgrid' or 'HSragrid', depending on the method of weighting.
  A 'HSrgrid' object includes the split, routed and weighted river network as first list element, and the
  'HSgrid' object  as the second list element. In A 'HSragrid' object, the first list element is the routed
  river network, the second element is a weighted area feature (catchment areas split at grid cell boundaries),
  and the third list element is the 'HSgrid' object.
}
\description{
Computes the weights used for each river segment when assigning values of an arbitrarily shaped polygon
  network. The function computes the weights either based on catchment areas(polygons), or line segments
  which intersect a grid cell. Weights are assigned so that the sum of weights for area or line features
  within a grid cell equals to 1.
}
\section{Details}{

  The river network needs to be connected, so that neighbouring river segments share a node, and all
  linestrings are cut at segment intersections (a "clean" network).

  Weights should be one of the following: "equal", "length", "strahler", "area", or a numeric vector
  which specifies the weight for each river segment. Equal, length, Strahler or user-defined weights are
  based on line segments. "equal" assigned grid cell value to all river segments within the cell, equally.
  "length" does the same, but weights are based on the length of river segment within the cell, so that
  longer segments get higher weights. "strahler" computes the strahler stream order for the river network
  and weights are based on that.

  If line-based weights are used, the river network is split at grid cell boundaries.

  "area" weights is used when catchment-based weighting is desired. It assigns cell value based on the area
  of catchment inside the grid cell. If no predefined basins or drainage directions are given, the function
  computes Voronoi polygons based on the river network input. If drainage direction raster is given the
  function computes catchment areas for each river segment in the input (however, at this stage the algorithm
  is implemented in R and is slow).

  Area of interest is optional. If provided, the river network will be clipped using it.
}

