% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_HSgrid.R
\name{create_HS}
\alias{create_HS}
\title{Constructs a \code{HS} object from given input}
\usage{
create_HS(zones, runoff, unit, zoneID = "zoneID", name = "runoff_1")
}
\arguments{
\item{zones}{An \code{sf POLYGON} object}

\item{runoff}{a table of runoff where rows are timesteps and columns
correspond to specific polygons in \code{grid}.}

\item{unit}{Unit of the runoff timeseries}

\item{zoneID}{Column name with unique IDs in \code{grid}.}

\item{name}{Name of the runoff timeseries. If multiple runoff timeseries
are added to the same \code{HS} object using \code{\link{add_HS}}, 
each timeseries must have a unique name, or else they are replaced.
@inheritParams raster_to_HS}
}
\value{
Returns a \code{HS} object with columns
  \itemize{
    \item zoneID: unique ID
    \item runoff_ts: a list column containing the runoff timeseries.
  }
}
\description{
Creates a \code{HS} from input where \code{grid} is a polygon grid
representing the spatial units of runoff, and /code{runoff} is a data.frame
containing column 'Date', and where column names match the grid IDs.
}
