% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSobs.methods.R
\name{create_HSobs}
\alias{create_HSobs}
\title{Constructor for \code{HSobs} object}
\usage{
create_HSobs(obs, riverIDs)
}
\arguments{
\item{obs}{a data.frame or a tibble with observations.}

\item{riverIDs}{A vector of riverID of the river segments of the columns
in obs.}
}
\value{
Returns a object of class \code{HSobs} which contains:
  \itemize{
    \item Observation timeseries
    \item riverIDs of the columns inthe observation timeseries
  }
}
\description{
Constructs an \code{HSobs} object from the input.
}
