% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_watershed.R
\name{delineate_basin}
\alias{delineate_basin}
\title{Delineate basin areas from multiple input points}
\usage{
delineate_basin(drain.dir, points, ID = "ID")
}
\arguments{
\item{drain.dir}{A RasterLayer object of drainage directions.}

\item{points}{An 'sf' point object, with locations of the basins to be delineated.}

\item{ID}{Name of the column in points with unique identifiers.}
}
\value{
Returns basins specific to the points given, so that downstream basins are cut when point location
  is met uptstream.
}
\description{
Delineate basin areas from multiple input points
}
