% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{DEM for examples}
\format{
RasterBrick / GeoTIFF
}
\usage{
dem
}
\description{
A 1 by 1 degree ALOS World 3D DEM resampled to 0.05 degree resolution. 
Original DEM by JAXA.
}
\details{
T. Tadono, H. Ishida, F. Oda, S. Naito, K. Minakawa, 
H. Iwamoto : Precise Global DEM Generation By ALOS PRISM, ISPRS Annals of the 
Photogrammetry, Remote Sensing and Spatial Information Sciences, Vol.II-4, 
pp.71-76, 2014.
}
\keyword{datasets}
