% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_downstream.R
\name{downstream}
\alias{downstream}
\title{Find all downstream segments from a specified river segment.}
\usage{
downstream(HSnetwork, ID, riverID = "riverID")
}
\arguments{
\item{HSnetwork}{a \code{HSnetwork} (routed river network) object obtained
with \code{\link{river_network}}.}

\item{ID}{The ID of the segment for which the downstream river network is 
returned.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to 
\code{"riverID"}.}
}
\value{
A \code{HSnetwork} object (a routed river network) consisting of the 
segments downstream from the segment specified.
}
\description{
Function returns the river network consisting of all river segments directly
downstream from the specified segment.
}
