% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_summary.R
\name{ensemble_summary}
\alias{ensemble_summary}
\title{Computes summaries from hydrostreamer::HS* objects}
\usage{
ensemble_summary(HS, summarise_over_all = TRUE, aggregate_monthly = FALSE,
  funs = c("min", "mean", "median", "max"), ...)
}
\arguments{
\item{HS}{Any HS* object, or a list of tables with column \code{Date}.}

\item{summarise_over_all}{Compute functions individually, or across all data.
Defaults to \code{TRUE}.}

\item{aggregate_monthly}{Summarise over the timestamps in data, or aggregate
results to the 12 months of the year.}

\item{funs}{Functions to evaluate. By default, computes \code{min, mean, median and 
max.}}

\item{...}{Additional arguments passed to \code{funs}.}
}
\value{
Returns the input HS* object, or a list, where runoff/downscaled/discharge
  is replaced with the computed summaries.
}
\description{
Allows easy computation of summaries across runoff datasets with user defined 
functions. The functions provided are run either individually for, or across, 
each runoff/downscaled/flow sets in an HS* object.
}
