% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instant_routing.R
\name{evaluate_instant_routing}
\alias{evaluate_instant_routing}
\title{Evaluates applicability of instantaneous flow routing}
\usage{
evaluate_instant_routing(HS, timestep, velocity = 1, verbose = FALSE)
}
\arguments{
\item{HS}{a \code{HS} object obtained passed through
\code{\link{river_network}}.}

\item{timestep}{Timestep length in \code{seconds}.}

\item{velocity}{Constant flow velocity in \code{meters per second}.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
A vector of the collected stats, or the input \code{HS} with a 
  specified column storing the vector.
}
\description{
Function to evaluate whether instantaneous flow routing may be used for
a certain river network and a specified timestep length.
}
