% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS_methods.R
\name{Extract_timeseries}
\alias{Extract_timeseries}
\alias{discharge}
\alias{runoff}
\alias{observations}
\alias{control}
\title{Convenience functions to extract timeseries from a HS* object}
\usage{
discharge(HS, riverID = NULL)

runoff(HS, riverID = NULL)

observations(HS, riverID = NULL)

control(HS, riverID = NULL)
}
\arguments{
\item{HS}{a \code{HS} object}

\item{riverID}{A vector of riverIDs for which to extract timeseries. If
\code{NULL} (default), extracts timeseries from all river segments.}
}
\value{
Returns a list of \code{tsibble}s with column Date and columns 
  named by riverID. Each element of the list is named by andcorrespond to a 
  timeseries in the timeseries column column specified by function name.
}
\description{
These functions can be used to quickly extract segment specific timeseries
(either discharge, runoff, observations or control) by riverID, or create
a table from specified riverIDs with date and timeseries of the river 
segments in columns.
}
