% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_gof.R
\name{flow_gof}
\alias{flow_gof}
\title{Computed goodness-of-fit statistics for discharge predictions
against observed streamflow}
\usage{
flow_gof(HSflow, HSobs)
}
\arguments{
\item{HSflow}{An \code{HSflow} object}

\item{HSobs}{An \code{HSobs} object containing observations at the
locations to optimise.}
}
\value{
Returns a list with gof-statistics for all the observations in
 every discharge prediction in \code{HSflow}.
}
\description{
Computed goodness-of-fit statistics using \code{\link[hydroGOF]{gof}}
for all the timeseries in the input \code{HSflow} object.
}
