% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_river.R
\name{flow_network}
\alias{flow_network}
\title{Generates neighbour information from a connected, directed graph without loops.}
\usage{
flow_network(river, ID = "ID")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{ID}{Name of the column containing unique segment ID's.}
}
\value{
Returns the river network with class 'HSrnet' with added columns:
\itemize{
  \item PREVIOUS: ID(s) of the previous river segment(s) as a list
  \item NEXT: ID of the segment where the river flows to
  \item DOWNSTREAM: ID(s) of all the river segments downstream for the current segment, as a list.
}
}
\description{
Generates neighbour information from a connected, directed graph without loops.
}
