% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rivers.R
\name{group_rivers}
\alias{group_rivers}
\title{Assign grouping factor id to each set of interconnected reaches}
\source{
\url{https://gis.stackexchange.com/a/310465}
}
\usage{
group_rivers(sf_riv_net, verbose = FALSE)
}
\arguments{
\item{sf_riv_net}{An sf LINESTRING object.}

\item{verbose}{Boolean. If TRUE, returns messages during processing.}
}
\value{
An sf LINESTRING object with one additional column.
}
\description{
This function receives a directed sf LINESTRING that describes a
  river network and returns an additional column with unique grouping id.
}
\details{
This function is based on igraph solution proposed by Barry
  Rowlingson (Follow URL below for details). The original solution used
  \emph{sf}, whereas the implemented solution uses \emph{geos}.
}
\examples{
library(dplyr)
library(sf)
set.seed(1234)
sf_riv_net = hydrostreamer::miera_rivers \%>\%
    dplyr::filter(
        !c(.data$riverID \%in\% sample(hydrostreamer::miera_rivers$riverID, 
                                               size = 1)))
sf_riv_net.grouped <- group_rivers(sf_riv_net,
                                            verbose = TRUE)

\dontrun{
library(ggplot2)
sf_riv_net.grouped \%>\%
 ggplot()+
 geom_sf(aes(geometry = geom, col = factor(river_group)),show.legend = FALSE) +
 theme_bw()
}
}
