% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSwrite.R
\name{HSwrite}
\alias{HSwrite}
\title{Writes any HS* object as a single geopackage where each input 
runoff is written as it's own layer.}
\usage{
HSwrite(x, filename, ...)
}
\arguments{
\item{x}{A \code{HSgrid}, \code{HSrunoff}, or a \code{HSflow} object.}

\item{filename}{Filename to write to.}

\item{...}{options passed to \code{\link[sf]{st_write}}.}
}
\description{
Writes any HS* object as a single geopackage where each input 
runoff is written as it's own layer.
}
