% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrostreamer-package.R
\docType{package}
\name{hydrostreamer-package}
\alias{hydrostreamer}
\alias{hydrostreamer-package}
\title{hydrostreamer: A Package for Interpolating Distributed Runoff Products on to Explicit River Segments}
\description{
This package provides functions to interpolate multiple distributed runoff data on to an explicitly represented river network, route runoff down the network, and optimise predictions against observed streamflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mkkallio.github.io/hydrostreamer/}
  \item Report bugs at \url{https://github.com/mkkallio/hydrostreamer/issues}
}

}
\author{
\strong{Maintainer}: Marko Kallio \email{marko@markokallio.fi} (\href{https://orcid.org/0000-0002-6917-7790}{ORCID})

Authors:
\itemize{
  \item Marko Kallio \email{marko@markokallio.fi} (\href{https://orcid.org/0000-0002-6917-7790}{ORCID})
  \item Vili Virkki \email{vili.virkki@aalto.fi} (\href{https://orcid.org/0000-0002-2603-3420}{ORCID})
}

}
\keyword{internal}
