% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrostreamer.R
\docType{package}
\name{hydrostreamer}
\alias{hydrostreamer}
\title{hydrostreamer: A package for interpolatinng distributed runoff products on to 
explicit river network}
\description{
Hydrostreamer provides functions to interpolate multiple distributed runoff 
data on to an explicitly represented river network, route runoff down the 
network, and optimise predictions against observed streamflow.
}
