% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_runoff.R
\name{interpolate_runoff}
\alias{interpolate_runoff}
\title{Estimate runoff in target features
  
Estimates runoff in either}
\usage{
interpolate_runoff(
  HS,
  river,
  basins = NULL,
  dasymetric = NULL,
  pycnophylactic = NULL,
  n = 20,
  intensive = TRUE,
  weights = NULL,
  aoi = NULL,
  riverID = "riverID",
  verbose = FALSE
)
}
\arguments{
\item{HS}{A 'HS' object, obtained with \code{\link{raster_to_HS}}.}

\item{river}{An 'sf' linestring feature representing a river network.}

\item{basins}{An 'sf' polygon object.}

\item{dasymetric}{Column name in \code{river} or in \code{basins} to be used 
as the ancillary information in dasymetric mapping. If \code{NULL} 
(default), no dasymetric mapping is performed. See details.}

\item{pycnophylactic}{Column in \code{HS} to be used as basis in 
pycnophylactic interpolation. If \code{NULL}, or if \code{is.null(basins)},
not performed. See details.}

\item{n}{Number of iterations when using pycnophylactic interpolation. 
Default \code{25}.}

\item{intensive}{Whether the pycnophylactic variable is intensive (density,
like runoff in mm), or not (in which case it is extensive, or counts like
runoff in volume).}

\item{weights}{Name of a column in \code{river} to be used directly as 
weights. Defaults to \code{NULL}. See Details.}

\item{aoi}{An area of interest ('sf' polygon object) used to intersect 
\code{river}. Ignored, if basins provided (in which case, aoi is the union
of \code{basins}). Defaults to \code{NULL}.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to 
\code{"riverID"}.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns a list object with class 'HSweights', containing the following 
  elements:
  \itemize{
    \item \code{river}. The supplied river network with routing information. 
      See \code{\link{river_network}} for details.
    \item \code{weights}. River network lines or catchment polygon network 
      which was used as the basis of weighting. See 
      \code{\link{compute_area_weights}} or \code{\link{compute_river_weights}} 
      for details.
    \item \code{HS}. The input HS object containing runoff information. 
}
}
\description{
Estimate runoff in target features
  
Estimates runoff in either
}
