% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miera_ERA5_land.R
\docType{data}
\name{miera_ERA5_land}
\alias{miera_ERA5_land}
\title{ERA5 land data for river Miera, Spain}
\format{
An RasterLayer: \describe{
  \item{file}{a ,RasterFile with 13 slots (brick)} ... }
}
\source{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\usage{
miera_ERA5_land
}
\description{
A dataset containing cropped ERA5 land runoff gridded timeseries data. See
data-raw for download code.
}
\references{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\keyword{datasets}
