% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miera_rivers_split.R
\docType{data}
\name{miera_rivers_split}
\alias{miera_rivers_split}
\title{Simplified river network for the river Miera, Spain (split)}
\format{
An sf object with 832 rows and 5 variables:
\describe{
  \item{riverID}{id for each river reach, a numeric}
  \item{old_riverID}{id for each river reach prior to split, a numeric}
  \item{river_group}{id for each river group, a numeric}
  \item{Maidment_velocity_ms}{Maidment velocity in m per s, a numeric}
  \item{geom}{simple feature geometry, sfc_LINESTRING sfc}
  ...
}
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-hydro/eu-hydro-river-network-database}
}
\usage{
miera_rivers_split
}
\description{
A simple feature dataset containing a subset of simplified EU-Hydro river
network for the river Miera, Spain. This dataset is derived from
hydrostreamer::miera_rivers. Here, network edges have been split to
facilitate the constant velocity routing algorithm with hourly climate grid
runoff product ER5-Land.
}
\references{
\url{https://github.com/ateucher/rmapshaper}
}
\keyword{datasets}
