% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSflow.methods.R
\name{plot_HSflow}
\alias{plot_HSflow}
\title{Plot HSflow}
\usage{
plot_HSflow(HSflow, HSobs)
}
\arguments{
\item{HSflow}{An \code{HSflow} object}

\item{HSobs}{An \code{HSobs} object}
}
\description{
Plots a timeseries of all predictions in HSflow.
}
