% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSoptim.methods.R
\name{plot_HSoptim}
\alias{plot_HSoptim}
\title{Plot HSoptim}
\usage{
plot_HSoptim(HSoptim, plot_type = c("timeseries", "residuals", "weights",
  "gof"), gof_stat = c("NSE", "RMSE", "ME", "R2", "KGE", "PBIAS.."),
  interactive = FALSE)
}
\arguments{
\item{HSoptim}{An \code{HSoptim} object}

\item{plot_type}{One of "timeseries", "residuals", "weights", or "gof". 
Defaults to "timeseries"}

\item{gof_stat}{Statistics to draw plot from. For options, see 
\code{\link[hydroGOF]{gof}}. Only used if \code{plot_type == "gof"}.}

\item{interactive}{Whether or not use package \code{plotly} to draw an
interactive plot, rather than static ggplot.}
}
\description{
Plots \code{HSoptim} object, obtained with \code{\link{optimise_point}}.
}
\details{
Four types of plots are available:
\itemize{
  \item "timeseries": plots a timeseries for each optimized station
  \item "residuals": plots residuals at each optimized station
  \item "weights": plots weights of runoff timeseries used to get the 
    optimized timeseries
  \item "gof": plots goodness-of-fit statistics for each station
}
}
