% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygrid_timeseries.R
\name{polygrid_timeseries}
\alias{polygrid_timeseries}
\title{Polygonizes a raster and adds cell values as attribute columns to an 'sf' polygon object.}
\usage{
polygrid_timeseries(brick, aoi = NULL, timesteps = NULL)
}
\arguments{
\item{brick}{A Raster* object.}

\item{aoi}{Area of interest as an 'sf' polygon object, 'SpatialPolygons', or 'SpatialPolygonsDataFrame'.
Optional, but recommended for all applications.}

\item{timesteps}{A logical vector with same length as there are layers in brick, or a numeric vector
specifying the layers to process. Optional. If not given, will process all raster layers.}
}
\value{
Returns an 'sf' polygon object, with each polygon representing a raster cell, and which has been
  cropped to the area of interest. Columns are ID, area of the polygon in m^2, and each timestep as separate
  column. Output has class 'HSgrid'.
}
\description{
Polygonizes a raster and adds cell values as attribute columns to an 'sf' polygon object.
}
