% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river}
\alias{river}
\title{Example river network}
\format{An 'sf' LINESTRING object with one attribute.
\describe{
  \item{ID}{unique river segment ID}
  \item{geometry}{'sfc' geometry column}
}}
\usage{
river
}
\description{
A river network of 71 segments used in the vignette of hydrostreamer.
}
\keyword{datasets}
