% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_outlets.R
\name{river_outlets}
\alias{river_outlets}
\title{Identify locations from where to start catchment delineation.}
\usage{
river_outlets(river, drain.dir)
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network.}

\item{drain.dir}{A RasterLayer object of drainage directions.}
}
\value{
Returns an 'sf' point object of identified raster cell centers. Includes all the attributes in the 
given \code{river} object.
}
\description{
The function identifies the coordinates of raster cells which are just upstream from river confluence
points identified from a river network layer. The cell just upstream from confluence allows delineation
of segment specific catchments.
}
\details{
The function logic is such that it looks at the last node-pair of each segment in the river network. The angle
of the last node pair is computed, and the raster cell from where the last node-pair originated is identified. 
Note that the logic of the function requires that the river network is derived from the same raster than the 
drainage directions input to the function.

Note: Drainage directions must be in coded 1-128:
\itemize{
  \item{1}{:East}
  \item{2}{:Southeast}
  \item{4}{:South}
  \item{8}{:Southwest}
  \item{16}{:West}
  \item{32}{:Northwest}
  \item{64}{:North}
  \item{128}{:Northeast}
}
}
\examples{
\dontrun{
# to be added
}
 
}
