% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_river_voronoi.R
\name{river_voronoi}
\alias{river_voronoi}
\title{Create segment-specific Voronoi polygons from a network.}
\usage{
river_voronoi(river, aoi, riverID = "ID", min = 10, tolerance = 0.001)
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{aoi}{An 'sf' polygon object, or an object which can be converted to 'sf' using st_as_sf(). Optional.}

\item{riverID}{A character string which specifies the name of the column containing unique identifier in
river network. Defaults to "ID".}

\item{min}{A numeric value. Removes segments which are shorter than the value in meters, in order to remove unnecessarily
short river segments. Defaults to 10 meters.}

\item{tolerance}{Radius of the buffer used to take difference. Defaults to 0.001 (in map units).}
}
\value{
Returns an 'sf' polygon object, with a column "ID" corresponding to the river segment IDs.
}
\description{
The function creates Voronoi polygons for each segment in a directed connected network, where the Voronoi polygons
join together at network segment intersections.
}
\section{Details}{

Creating the segment Voronoi polygons is done in the following steps:
\enumerate{
  \item Extract segment end nodes.
  \item Create a buffer of the end nodes, and take difference between the river network and the buffer layer.
  \item Extract all nodes of the river network, and create Voronoi polygons for all nodes.
  \item Dissolve the individual Voronoi using river segment ID.
  \item Clip the polygons to the area of interest.
}

Thus, the accuracy of the final segment Voronoi diagram is depending on the resolution of nodes in the river network.
Consider densifying geometry e.g. with sf::st_segmentize function for higher accuracy.
}

