% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runoff}
\alias{runoff}
\title{Runoff timeseries for examples}
\format{RasterLayer / GeoTIFF}
\usage{
runoff
}
\description{
1 by 1 degree runoff timeseries in the same area as the DEM. Runoff is sourced 
from the Linear Optimal Runoff Aggregate (LORA) at 0.5 degree resolution.
The unit of runoff is mm/s (kg/m2/s), and are provided with a monthly timestep.
}
\details{
Hobeichi, S., Abramowitz, G., Evans, J., and Beck, H. E.: Linear Optimal 
Runoff Aggregate (LORA): a global gridded synthesis runoff product, 
Hydrol. Earth Syst. Sci., 23, 851-870, https://doi.org/10.5194/hess-23-851-2019, 
2019.
}
\keyword{datasets}
