% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_data.R
\docType{data}
\name{runoff_data}
\alias{runoff_data}
\title{LORA data}
\format{
A RasterBrick: \describe{
  \item{file}{a ,RasterBrick with dimensions 2, 2, 4, 396  (nrow, ncol, ncell, nlayers)} ... }
}
\source{
\url{https://doi.org/10.5194/hess-23-851-2019}
}
\usage{
runoff_data
}
\description{
A dataset containing cropped LORA runoff gridded timeseries data (Hobeichi et
al 2019). Resolution .5 degrees. Time series 1980-01-01 20 2012-12-01 monthly
timestep.
}
\references{
\url{https://doi.org/10.5194/hess-23-851-2019}
}
\keyword{datasets}
