% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_DEM.R
\docType{data}
\name{shannon_DEM}
\alias{shannon_DEM}
\title{Digital Elevation Model data for the river Shannon, Ireland}
\format{
An RasterLayer: \describe{
  \item{file}{a ,RasterFile with 13 slots} ... }
}
\source{
\url{https://registry.opendata.aws/terrain-tiles}
}
\usage{
shannon_DEM
}
\description{
A dataset containing cropped Mapzen Terrain Tiles from Amazon Web Services
accessed using elevatr R package. Data were downloaded at zoom level 10. See
data-raw for download code.
}
\references{
\url{https://github.com/jhollist/elevatr}
}
\keyword{datasets}
