% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_ERA5_land.R
\docType{data}
\name{shannon_ERA5_land}
\alias{shannon_ERA5_land}
\title{ERA5 land data for river Shannon, Ireland}
\format{
An RasterLayer: \describe{
  \item{file}{a ,RasterFile with 13 slots (brick)} ... }
}
\source{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\usage{
shannon_ERA5_land
}
\description{
A dataset containing cropped ERA5 land runoff gridded timeseries data. See
data-raw for download code.
}
\references{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\keyword{datasets}
