% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_ERA5_land_daily.R
\docType{data}
\name{shannon_ERA5_land_daily}
\alias{shannon_ERA5_land_daily}
\title{ERA5 land data for river Shannon, Ireland aggregated to daily time step}
\format{
An RasterLayer: \describe{
  \item{file}{a ,RasterFile with 13 slots (brick)} ... }
}
\source{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\usage{
shannon_ERA5_land_daily
}
\description{
A dataset containing cropped ERA5 land surface runoff gridded timeseries
data. See data-raw for download code.
}
\references{
\url{https://doi.org/10.24381/cds.e2161bac}
}
\keyword{datasets}
