% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_GRUN.R
\docType{data}
\name{shannon_GRUN}
\alias{shannon_GRUN}
\title{GRUN data for river Shannon, Ireland}
\format{
An RasterLayer: \describe{
  \item{file}{a ,RasterFile with 13 slots (brick)} ... }
}
\source{
\url{http://hdl.handle.net/20.500.11850/324386}
}
\usage{
shannon_GRUN
}
\description{
A dataset containing cropped GRUN runoff gridded timeseries data. See
data-raw for download code.
}
\references{
\url{https://figshare.com/articles/dataset/GRUN_Global_Runoff_Reconstruction/9228176}
}
\keyword{datasets}
