% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_rivers.R
\docType{data}
\name{shannon_rivers}
\alias{shannon_rivers}
\title{Simplified river network for the river Shannon, Ireland}
\format{
An sf object with 2206 rows and 7 variables:
\describe{
  \item{riverID}{id for each river reach, a numeric}
  \item{river_group}{id for each river group, a numeric}
  \item{region_id}{AMBER region code, a numeric}
  \item{lake_id}{id for lake nitersecting river reach, an integer}
  \item{gmean_m}{Elevation above sea level in m, a numeric}
  \item{grad}{Gradient from start to end node of reach, a numeric}
  \item{geom}{simple feature geometry, sfc_LINESTRING sfc}
  ...
}
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-hydro/eu-hydro-river-network-database}
}
\usage{
shannon_rivers
}
\description{
A simple feature dataset containing a subset of simplified EU-Hydro river
network for the river Shannon, Ireland. This LINESTRING was simplified to 5%
original detail using the rmapshaper package, which preserves topologies.
}
\references{
\url{https://github.com/ateucher/rmapshaper}
}
\keyword{datasets}
