% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_network_at_bifurcations.R
\name{split_network_at_bifurcations}
\alias{split_network_at_bifurcations}
\title{Split a HS network into spatially distinct chunks for iterative runoff
routing of complex anabranching digital river networks}
\usage{
split_network_at_bifurcations(
  sf_river_network,
  riverID = "riverID",
  verbose = FALSE
)
}
\arguments{
\item{sf_river_network}{A LINESTRING simple feature collection.}

\item{riverID}{unique edge identifier.}

\item{verbose}{Boolean. Print messages.}
}
\value{
Returns the river network with an additional three columns denoting
  the the spatially distinct subgroups "river_group_new", reach specific
  topological order "topo_order", a new NEXT column that replaces values for
  the most downstream segment in each new topologically ordered subgroup with
  -9999 for compatibility with _complex routing methods. The original NEXT
  column is renamed to "NEXT_stored".
}
\description{
A method for splitting vector river networks into spatially distinct chunks
for more efficient runoff routing of anabranching networks than simply
running hydrostreamer::accumulate_runoff().
}
\details{
This function is designed to topologically order a digital
  river network (LINESTRING sfc), identify each downstream bifurcation point
  in the network and split the network at these points using a small buffer
  (1cm). Then the function identifies the resulting spatially distinct
  groups maintaining topological order.
}
\examples{
# Miera example
library(dplyr)

 miera = hydrostreamer::miera_rivers \%>\%
     dplyr::mutate("default_order" = dplyr::row_number()) \%>\%
     hydrostreamer::river_network(riverID = "riverID",
                                  verbose = TRUE)

 # miera \%>\%
 # ggplot()+
 #    geom_sf(aes(geometry = geom, col = default_order))

 miera_split_at_bifurcations <-
     split_network_at_bifurcations(miera,
   riverID = "riverID",
   verbose = TRUE)
   
\dontrun{
library(ggplot2)

# individual edges
miera_split_at_bifurcations \%>\%
    ggplot()+
    geom_sf(aes(geometry = geom, col = topo_order)) +
    scale_color_gradientn(colours = c("blue","white","red"))

# groups
miera_split_at_bifurcations \%>\%
    ggplot()+
    geom_sf(aes(geometry = geom, col = river_group_new)) +
    scale_color_gradientn(colours = c("blue","white","red"))
}
}
