% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_river.R
\name{split_river_with_grid}
\alias{split_river_with_grid}
\title{Split a linestrings at polygon boundaries.}
\usage{
split_river_with_grid(river, grid)
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{grid}{A 'HSgrid' object. Required.}
}
\value{
Returns an 'sf' linestring object which has been split at the polygon (grid) boundaries,
  which has an added attribute column 'gridID' specifying the ID of the grid that contains the
  line segment.
}
\description{
Split a linestrings at polygon boundaries.
}
