% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{switch_grid}
\alias{switch_grid}
\title{Switch the runoff grid in HSweights object}
\usage{
switch_grid(HSweights, grid, gridID = "gridID")
}
\arguments{
\item{HSweights}{A \code{HSweights} object obtained with \code{\link{compute_weights}}.}

\item{grid}{A 'HSgrid' object, obtained with \code{\link{polygrid_timeseries}}.}

\item{gridID}{Column name in \code{grid} containing unique IDs.}
}
\value{
\code{HSweights} object with a replaced \code{HSgrid} element.
}
\description{
The functions changes the runoff grid in a \code{HSweights} object to another. Some simple tests are made to
the replacing \code{HSgrid} object to ensure that they are equivalent to the original one. The main use of this
function is to make it easy to compare multiple runoff products without the need to compute weights everytime.
}
\details{
The replacing grid must have the same spatial objects as in the one being replaced. This is because the weights
are calculated using spatial relatioships between the river segments (or segment specific catchments) and the
areal units in runoff time series. If the runoff unit areas change, so does the weighting, and in this case 
weights need to be recomputed using \code{\link{compute_weights}}.
}
