% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_flow.R
\name{accumulate_flow}
\alias{accumulate_flow}
\title{Apply river routing based on selected routing method.}
\usage{
accumulate_flow(HSrunoff, method = "instant", verbose = FALSE)
}
\arguments{
\item{HSrunoff}{A 'HSrunoff' object obtained by \code{\link{compute_segment_runoff}}}

\item{method}{Character string specifying the method to be used.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns the river network (without routing information) where runoff timeseries is switched to 
routed and accumulated flow.
}
\description{
Apply river routing using any of the provided methods using a monthly timestep. The function takes a 
\code{HSrunoff} object as an input.
}
\details{
Currently only instantaneous flow downstream is implemented, specified by value "simple" in method input.
}
