% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff.R
\name{accumulate_runoff}
\alias{accumulate_runoff}
\title{Apply river routing}
\usage{
accumulate_runoff(HS, method = c("instant", "constant"), ..., verbose = FALSE)
}
\arguments{
\item{HS}{A 'HS' object obtained by \code{\link{interpolate_runoff}}}

\item{method}{Character string specifying the method to be used.}

\item{...}{Arguments passed to the routing algorithm, and to 
\code{\link{river_network}}, if it has not been run already.}

\item{verbose}{Whether or not to print progress information. Defaults to 
\code{FALSE}.}
}
\value{
Returns the input object \code{HS}) with an added list column
  \code{discharge_ts} containing routed discharge estimates for each river
   segment.
}
\description{
Apply river routing using any of the provided methods. The function takes 
\code{HS} object with runoff timeseries as an input. See details.
}
\details{
There are currently two routing algorithms implemented: 
\code{'instant'}aneous flow, where all runoff is routed through the entire 
river network at every timestep. There is no lag in instantaneous routing, 
which  means it may not be reasonable for large river networks. 
\code{'constant'} velocity river routing routes runoff down the river network
with a constant, user specified velocity. See further details from 
the documentation of each method:
\itemize{
  \item \code{\link{accumulate_runoff_instant}}
  \item \code{\link{accumulate_runoff_constant}}
}
}
