% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_runoff_constant.R
\name{accumulate_runoff_constant}
\alias{accumulate_runoff_constant}
\title{Apply simple lag river routing}
\usage{
accumulate_runoff_constant(HS, velocity = 1, verbose = FALSE)
}
\arguments{
\item{HS}{A \code{HS} object.}

\item{velocity}{Flow velocity applied to compute parameter x. Can be a 
constant, or a vector of flow velocity at each unique river segments.
Flow velocity defaults at 1.}

\item{verbose}{Whether or not to print progress information. Defaults to 
\code{FALSE}.}
}
\value{
Returns the input object \code{HS}) with an added list column
  \code{discharge_ts} containing routed discharge estimates for each river
   segment.
}
\description{
Implements a simple lag-based routing algorithm which can be used with
arbitrary intervals between runoff units. The river flow is 'slowed' down by
computing how much water the river segment has at any one time. This 'storage'
is added to the next timestep.
}
