% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_HSgrid.R
\name{add_HSgrid}
\alias{add_HSgrid}
\title{Combine \code{HSgrid} objects or addnew runoff input from raster.}
\usage{
add_HSgrid(HSgrid, from = NULL, raster = NULL, date = NULL,
  timestep = NULL, aoi = NULL, name = NULL)
}
\arguments{
\item{HSgrid}{an existing \code{HSgrid} object.}

\item{from}{A \code{HSgrid} object to add from. Optional.}

\item{raster}{A Raster* object.}

\item{timestep}{Length of timestep in raster, e.g. \code{"hour"}, 
\code{"day"}, \code{"month"}}

\item{aoi}{Area of interest as an 'sf' polygon object, 
'SpatialPolygons', or 'SpatialPolygonsDataFrame'. Optional, but 
recommended for all applications.}
}
\value{
Returns the input \code{HSgrid} object with added runoff.
}
\description{
Combines the runoff lists from two HSgrid objects, or adds a new
runoff table from a raster. \emph{Warning: The geometries of the
combined grids must be identical.}
}
