% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observations.R
\name{add_observations}
\alias{add_observations}
\title{Adds observation timeseries to a 'HS' object}
\usage{
add_observations(HS, timeseries, riverIDs, station_names = NULL)
}
\arguments{
\item{HS}{An \code{HS} object.}

\item{timeseries}{a data.frame with observations. Must contain column 
\code{Date}.}

\item{riverIDs}{A vector of riverID of the river segments of the columns
in timeseries.}

\item{station_names}{a vector of names for the stations in \code{timeseries}.
If not provided, station names are obtained from column names of
\code{timeseries}.}
}
\value{
Returns the \code{HS} object with added list column 
  \code{observation_ts} containing a timeseries, and column 
  \code{observation_station} containing the name of observation stations.
}
\description{
Adds observation timeseries to a HS object. This is needed in order to 
evaluate performance of downscaled timeseries, or to perform data assimilation
combining several downscaled times.
}
