% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_monthly_runoff.R
\name{average_monthly_runoff}
\alias{average_monthly_runoff}
\title{Monthly average of a runoff timeseries. Expects the timeseries to start from January and end in December.}
\usage{
average_monthly_runoff(x)
}
\arguments{
\item{x}{A 'HSgrid' object obtained with \code{\link{polygrid_timeseries}}, or a 'RasterStack', or a 'RasterBrick'.}
}
\value{
Returns either a 'HSgrid' object with monthly summary of the timeseries, or a 'RasterBrick',
  depending on the class of input.
}
\description{
Monthly average of a runoff timeseries. Expects the timeseries to start from January and end in December.
}
