% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runoff.R
\name{combine_runoff}
\alias{combine_runoff}
\title{Combines runoff in HS* objects}
\usage{
combine_runoff(HS, weights = NULL)
}
\arguments{
\item{HS}{An \code{HS*} object}

\item{weights}{A vector of weights with the same length as the number of
runoff timeseries in \code{HS}, or a list of weight vectors.}
}
\value{
Returns the input \code{HS} object where runoff timeseries is 
  replaced by the the combined timeseries'.
}
\description{
The function provides the ability to combine runoff from \code{HSgrid}, 
\code{HSweights}, \code{HSrunoff}, or \code{HSflow} using user provided
weights. Function can process several input weight sets at once.
}
\details{
If a named list of weights is provided to the function, their names will be 
preserved in the output. Otherwise the combinations are named in a running
order.

Note that if the timeseries of dates is different in each runoff timeseries
in the input, the output includes only those dates which are present in 
every input timeseries.
}
\examples{
\dontrun{
  weights <- c(0.2,0.2,0.3,0.3,0)
  combined <- combine_runoff(HSgrid, weights)
  
  listweights <- list(weights1 = c(0.2,0.2,0.3,0.3,0),
                  weights2 = c(0.1,0.2,0.3,0.2,0.2),
                  thirdw = c(0,0,0,0.5,0.5))
 combined2 <- combine_runoff(HSgrid, listweights)
}


}
