% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_area_weights}
\alias{compute_area_weights}
\title{Assign weights to area features within grid cells.}
\usage{
compute_area_weights(basins, grid)
}
\arguments{
\item{basins}{An 'sf' polygon feature.}

\item{grid}{A 'HSgrid' object. Required.}
}
\value{
Returns an 'sf' polygon feature (which is a union of basins, and grid) with an attribute column
  'weights', which is the area of the polygon feature inside a grid cell, divided by the total area
  of the grid cell.
}
\description{
Assign weights to area features within grid cells.
}
