% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_area_weights.R
\name{compute_area_weights}
\alias{compute_area_weights}
\title{Compute weights for river catchment areas within the runoff area 
features.}
\usage{
compute_area_weights(basins, HSgrid, seg_weights = NULL,
  riverID = "riverID", gridID = "gridID")
}
\arguments{
\item{basins}{An 'sf' polygon feature specifying the river segment 
specific catchments.}

\item{HSgrid}{A 'HSgrid' object, obtained with \code{\link{raster_to_HSgrid}}.}

\item{riverID}{Column in \code{basins} containing unique IDs.}

\item{gridID}{Column in  \code{HSgrid} with unique IDs.}
}
\value{
Returns an 'sf' polygon feature (a union of basins, and HSgrid) 
  with added attributes (columns):
  \itemize{
    \item \emph{ID}. Unique ID of the feature.
    \item \emph{riverID}. ID of the river segment each sub-catchment is 
      associated to.
    \item \emph{gridID}. ID of the runoff unit the sub-catchment 
      is contained in.
    \item \emph{weights}. Weights computed for each sub-catchment.
    \item \emph{b_area_m2}. Area of the sub-catchment (basin) in 
      \eqn{m^2}.
    \item \emph{g_area_m2}. Area of the runoff unit sub-catchment is 
      contained in. In \eqn{m^2}.
}
}
\description{
Computes weights for each individual river segment specific catchments 
falling in the areal units of the runoff \emph{HSgrid}. Function first 
takes a union between \emph{basins} and \emph{HSgrid} (creating new 
catchment units which fall inside only one runoff unit), and calculating 
the area for each individual catchment unit. The weight is assigned by 
dividing the area of sub-catchment with the area of runoff unit.
This function is called by \code{\link{compute_HSweights}}.
}
