% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow.R
\name{compute_segment_runoff}
\alias{compute_segment_runoff}
\title{Computes specific runoff generated in the river segment based on either polygon or area weighting.}
\usage{
compute_segment_runoff(x, timesteps = NULL)
}
\arguments{
\item{x}{An object of class 'HSrgrid' or 'HSragrid', obtained with compute_weights().}

\item{timesteps}{A logical vector of the length of timesteps in x, or a numeric vector specifying which timesteps to process. If no timesteps are given, all will be processed.}
}
\value{
River network ('sf' linestring object) with river segment specific runoff for all timesteps processed as attribute columns. Object class 'HSrunoff'.
}
\description{
Computes specific runoff generated in the river segment based on either polygon or area weighting.
}
\examples{
\dontrun{
compute_segment_runoff(voronoi_weigths)
compute_segment_runoff(voronoi_weigths, timesteps=1:12)
}
}
