% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_HSgrid.R
\name{create_HSgrid}
\alias{create_HSgrid}
\title{Constructs a \code{HSgrid} object from given input}
\usage{
create_HSgrid(grid, runoff, name = "runoff")
}
\arguments{
\item{grid}{An \code{sf POLYGON} object}

\item{runoff}{a table of runoff where rows are timesteps and columns
correspond to specific polygons in \code{grid}.}

\item{name}{Name of the runoff timeseries. If multiple runoff timeseries
are added to the same \code{HSgrid} object using \code{\link{add_HSgrid}}, 
each timeseries must have a unique name, or else they are replaced.}
}
\value{
Returns a \code{HSgrid} object which is a list of:
  \itemize{
    \item grid: polygon grid with areal units of runoff
    \item runoff: a list of runoff values for each polygon.
  }
}
\description{
Creates a \code{HSgrid} from input where grid is a polygon grid
representing the units of runoff. It must include a column 'gridID'
with unique IDs. Runoff must contain column 'Date', and value for 
runoff in columns, where the column name is the ID of the corresponding 
polygon in grid.
}
