% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_HSweights.R
\name{create_HSweights}
\alias{create_HSweights}
\title{Creates a \code{HSweights} object}
\usage{
create_HSweights(river, weights, grid, riverID = "riverID")
}
\arguments{
\item{river}{A river network as an \code{sf LINESTGRING} object}

\item{weights}{Weights with riverID, gridID and weight information.}

\item{grid}{A \code{HSgrid} object.}
}
\value{
Returns a \code{HSweights} object, which is a list of three 
  components:
  \itemize{
    \item river: River network with routing information to be used in
      downscaling step.
    \item weights: Weighting information for the downscaling step.
    \item grid: \code{HSgrid} object with runoff timeseries information.
  }
}
\description{
Creates a list object with class \code{HSweights}. The river object
can be obtained with \code{\link{river_network}}, weights with
\code{\link{compute_river_weights}} or \code{\link{compute_area_weights}},
and grid with \code{\link{raster_to_HSgrid}}.
}
