% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.R
\name{delineate_basin}
\alias{delineate_basin}
\title{Delineate catchments from multiple input points}
\usage{
delineate_basin(outlets, drain.dir, riverID = "riverID", verbose = FALSE)
}
\arguments{
\item{outlets}{An 'sf' point object, with locations of the catchment outlets for delineation. Obtained
e.g. with \code{\link{river_outlets}}}

\item{drain.dir}{A RasterLayer object of drainage directions.}

\item{riverID}{Name of the column in outlets with unique identifiers.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns polygonized catchment specific to the outlet points given. The output is 'sf' POLYGON class
with attributes:
\itemize{
  \item riverID: The ID given in \code{outlets}.
  \item ncells: The number of raster cells the catchment consists of.
}
}
\description{
This algorithm delineates multiple catchments from input point locations and a drainage direction layer.
Catchments are delineated upstream from the point given, until another given outlet point is met. The result
is a set of catchments which are unique to the points given.
}
\details{
Note that this function is currently implemented using R alone, which means that the for large areas of interest,
computation time may take long. Reading the drainage direction raster entirely to memory (using 
\code{\link[raster]{readAll}} from the \code{raster} package) speeds up the computation#' considerably, and is strongly 
suggested.
}
\examples{
\dontrun{
# to be added
}

}
