% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_runoff.R
\name{downscale_runoff}
\alias{downscale_runoff}
\title{Computes specific runoff generated in the river segment.}
\usage{
downscale_runoff(
  HSweights,
  pycno = FALSE,
  n = 10,
  dasy = NULL,
  verbose = FALSE
)
}
\arguments{
\item{HSweights}{An object of class 'HSweights', obtained with 
\code{compute_HSweights}, or constructed with function
\code{create_HSweights}.}

\item{pycno}{If \code{TRUE}, ignore weights and perform pycnophylactic
interpolation for every timestep. EXPERIMENTAL AND 2-3 ORDERS OF MAGNITUDE
SLOWER THAN USING WEIGHTS. Default FALSE.}

\item{n}{Iterations in the pycnophylactic interpolation.}

\item{dasy}{name of the dasymetric variable in \code{HSweights$weights}.
Experimental, used in pycnophylactic interpolation.}

\item{verbose}{Print progress indication or not.}
}
\value{
The routed river network object with class \code{HS}, which has been 
  enhanced with a runoff timeseries (list column \code{runoff_ts}. Runoff is 
  given in \eqn{m^3/s}.
}
\description{
Computes downscaled, river segment specific, runoff using 
the given \code{HSweights} object.
}
