% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_runoff.R
\name{downscale_runoff}
\alias{downscale_runoff}
\title{Computes specific runoff generated in the river segment.}
\usage{
downscale_runoff(HSweights, rID = "riverID", wID = "riverID",
  unit = "mm/s")
}
\arguments{
\item{HSweights}{An object of class 'HSweights', obtained with 
\code{compute_HSweights}, or constructed with function
\code{create_HSweights}.}

\item{rID}{Name of the column in river with unique IDs.}

\item{wID}{Name of the column in weights with IDs corresponding to rID.}

\item{unit}{Unit of runoff. Can be either "mm/s", or "m3/s". 
Defaults to mm/s (~ equivalent to kg/m2/s).}
}
\value{
The routed river network object which has been enhanced with 
  a runoff timeseries. Runoff is given in \eqn{m^3/s}.
}
\description{
Computes downscaled, unrouted, river segment specific runoff using 
the given \code{HSweights} object.
}
\examples{
\dontrun{
library(raster)
library(hydrostreamer)

# load data
data(river)
data(basin)
runoff <- brick(system.file("extdata", "runoff.tif", package = "hydrostreamer"))

# create HSgrid
grid <- polygrid_timeseries(grid, aoi=basin)

# weights based on length of river segments
len <- compute_weights(river, grid, "length", aoi=basin, riverID="ID")

# compute segment specific runoff
runoff <- compute_segment_runoff(len, verbose=TRUE)
}

}
