% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_gof.R
\name{flow_gof}
\alias{flow_gof}
\title{Computed goodness-of-fit statistics for discharge predictions
against observed streamflow}
\usage{
flow_gof(HS)
}
\arguments{
\item{HS}{A \code{HS} object with column \code{discharge_ts}}
}
\value{
Returns a list with gof-statistics for all the observations in
 every discharge prediction in \code{HS}.
}
\description{
Computed goodness-of-fit statistics using \code{\link[hydroGOF]{gof}}
for all the timeseries in the input \code{HS} object.
}
\details{
By default, goodness-of-fit measures are computed for all timeseries in 
the input data.
}
