% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrostreamer.R
\docType{package}
\name{hydrostreamer}
\alias{hydrostreamer}
\alias{hydrostreamer-package}
\title{hydrostreamer: A package for assigning runoff from land surface models to explicit river segments.}
\description{
hydrostreamer: A package for assigning runoff from land surface models to explicit river segments.
}
