% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrostreamer.R
\docType{package}
\name{hydrostreamer}
\alias{hydrostreamer}
\title{hydrostreamer: A package for downscaling distributed runoff products on to 
explicit river segments.}
\description{
hydrostreamer provides functions to downscale distributed runoff 
data on to an explicitly represented river network. Downscaling is done by the 
spatial relationship between an areal unit of runoff and an overlaid river 
network. Value of the runoff unit is divided among intersecting river segments
using weighted interpolation. hydrostreamer provides several methods for the 
assignment. Simple river routing algorithms are also provided to estimate 
discharge at arbitrary segment.
}
