% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_runoff.R
\name{interpolate_runoff}
\alias{interpolate_runoff}
\title{Estimate runoff in river reaches 
  
Estimates runoff in individual segments of a river network represented as 
either \code{sf LINESTRING} or \code{sf POLYGON} objects. Allows use of 
simple Area Weighted Interpolation, Area-to-line interpolation, Dasymetric
Mapping, Pycnophylactic Interpolation, or combined Pycnophylactic-Dasymetric
Interpolation.}
\usage{
interpolate_runoff(
  HS,
  river,
  basins = NULL,
  dasymetric = NULL,
  pycnophylactic = NULL,
  n = 10,
  intensive = TRUE,
  weights = NULL,
  aoi = NULL,
  riverID = "riverID",
  verbose = FALSE
)
}
\arguments{
\item{HS}{A 'HS' object, obtained with \code{\link{raster_to_HS}} or 
\code{\link{create_HS}}.}

\item{river}{An 'sf' linestring or polygon feature representing a river 
network}

\item{basins}{An 'sf' polygon object with corresponding catchments, if river
input is a linestring. Optional.}

\item{dasymetric}{Column name in \code{river} or in \code{basins} to be used 
as the ancillary information in dasymetric mapping. If \code{NULL} 
(default), no dasymetric mapping is performed.}

\item{pycnophylactic}{Column in \code{HS} to be used as basis in 
pycnophylactic interpolation. If \code{NULL}, or if \code{is.null(basins)},
not performed.}

\item{n}{Number of iterations when using pycnophylactic interpolation. 
Default \code{10}.}

\item{intensive}{Whether the pycnophylactic variable is intensive (density,
like runoff in mm), or not (in which case it is extensive, or counts like
runoff in volume).}

\item{weights}{Name of a column in \code{river} to be used directly as 
weights. Defaults to \code{NULL}. See Details.}

\item{aoi}{An area of interest ('sf' polygon object) used to intersect 
\code{river}. Ignored, if basins provided (in which case, aoi is the union
of \code{basins}). Defaults to \code{NULL}.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to 
\code{"riverID"}.}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns a \code{HS} object, containing the following 
  elements:
  \itemize{
    \item \code{riverID}. Unique identifier 
    \item \code{runoff_ts}. Interpolated runoff timeseries at each river 
      reach.
}
}
\description{
The river network can be any of three options: 1) a river network consisting
of LINESTRINGs, in which case Area-to-Line interpolation is used. 2) A 
network of POLYGONs, in which case Areal Interpolation (AWI, DM, or PP) 
is used. 3) both LINESTRING river network, and a POLYGON network describing 
the catchments of each individual segment of the LINESTRING network. In this 
case, Areal Interpolation is used. The third option is useful for routing 
purposes. If only catchments are given, only instantaneous routing can be 
used, because river line length cannot be derived from polygon areas.
}
\details{
The selection of areal interpolation method is determined by the input to the
function. If \code{dasymetric} and \code{pycnophylactic} are NULL (default),
simple Area Weighted Interpolation or Area-to-Line Interpolation is used. If
\code{dasymetric} variable is given, Dasymetric Mapping is used instead. 
If \code{pycnophylactic} is given, the function performs Pycnophylactic 
Interpolation. And if both \code{dasymetric} and \code{pycnophylactic} are 
given, the function first performs Pycnophylactic Interpolation, which is 
further refined by the given dasymetric variable. Note that pycnophylactic
interpolation is only available for polygon networks. Catchments for 
linestring networks can be estimated using \code{\link{river_voronoi}},
allowing pycnophylactic interpolation for line networks.
}
