% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_point.R
\name{optimise_point}
\alias{optimise_point}
\title{Computes optimal estimate from multiple input discharge timeseries}
\usage{
optimise_point(HSflow, HSobs, optim_method = "CLS", train = 0.5,
  bias_correction = FALSE)
}
\arguments{
\item{HSflow}{An \code{HSflow} object}

\item{HSobs}{An \code{HSobs} object containing observations at the
locations to optimise.}

\item{optim_method}{Method used to optimise. Default uses constrained
linear regression. See details.}

\item{train}{The share of timeseries used for training period.}

\item{bias_correction}{Apply (constant) bias correction. Defaults to FALSE.}
}
\value{
Returns an object of class \code{HSoptim}, which is a list of
results from observation stations. Each list item contains:
  \itemize{
    \item Observations: Supllied observation timeseries.
    \item Forecast_train: Forecasted timeseries at training period.
    \item Forecast_test: Forecasted timeseries at testing period.
    \item Forecast_train_test: Forecasted timeseries with training and
      testing combined
    \item Method: Used method for forecast combination.
    \item Forecast_weights: Vector of weights - how the input discharge 
      timeseries were combined.
    \item Intercept: Intercept from the combination. \code{NA}, if not
      applicable to the Method.
    \item Goodness_of_fit. Goodness of fit of the forecast combination
      obtained using \code{\link[hydroGOF]{gof}}.
}
}
\description{
Function performs timeseries optimisation of discharge estimates against
observed streamflow timeseries at specific river segments.
}
\details{
Optimisation of the input timeseries against observations makes use of 
\code{forecastComb} package, with three options currently supported;
\code{"CLS"} (constrained linear regression), \code{"OLS"} (ordinary
linear regression), and \code{"factorCLS"} (constrained linear regression
in factorized form). Default is constrained linear regression, which 
constraints the coefficients to add to unity, and all weights are between
0, and 1. CLS also forces Intercept to 0. In some cases, CLS fails to 
find a unique solution. In those cases, you can try the factorized
CLS.

If bias correction is set to TRUE, a constant bias correction is applied
to the entire timeseries so that mean error in the training period is 0.
}
