% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_point.R
\name{optimise_point}
\alias{optimise_point}
\title{Computes optimal estimate at a specific river segment}
\usage{
optimise_point(HS, optim_method = "CLS", combination = "ts",
  sampling = "series", train = 0.5, bias_correction = FALSE)
}
\arguments{
\item{HS}{An \code{HS} object with observation_ts and discharge_ts}

\item{optim_method}{Method used to optimise. Default uses constrained
linear regression. See details.}

\item{combination}{Whether to do the forecast combination for the entire
timeseries, or each month of the year individually. Accepts 
\code{"timeseries"}, \code{"ts"}, \code{"monthly"} or \code{"mon"}.}

\item{sampling}{How to sample training and testing periods. \code{"series"}
for training serially from beginning, or \code{"random"} for a random
sample for both training and testing periods.}

\item{train}{The share of timeseries used for training period.}

\item{bias_correction}{Apply bias correction. Defaults to FALSE.}
}
\value{
Returns an object of class \code{HSoptim}, which is a list of
results from observation stations. Each list item contains:
  \itemize{
    \item Observations: Supllied observation timeseries.
    \item Forecast_train: Forecasted timeseries at training period.
    \item Forecast_test: Forecasted timeseries at testing period.
    \item Forecast_train_test: Forecasted timeseries with training and
      testing combined
    \item Method: Used method for forecast combination.
    \item Forecast_weights: Vector of weights - how the input discharge 
      timeseries were combined.
    \item Intercept: Intercept from the combination. \code{NA}, if not
      applicable to the Method.
    \item Goodness_of_fit. Goodness of fit of the forecast combination
      obtained using \code{\link[hydroGOF]{gof}}.
}
}
\description{
Function performs data assimilation by combining timeseries of downscaled 
discharge estimates against observed streamflow timeseries at all river 
segments with observations.
}
\details{
Optimisation of the input timeseries against observations makes use of 
\code{forecastComb} package, with three options currently supported;
\code{"CLS"} (constrained linear regression), \code{"OLS"} (ordinary
linear regression), and \code{"factorCLS"} (constrained linear regression
in factorized form). Default is constrained linear regression, which 
constraints the coefficients to add to unity, and all weights are between
0, and 1. CLS also forces Intercept to 0. In some cases, CLS fails to 
find a unique solution. In those cases, you can try the factorized
CLS.

If bias correction is set to TRUE, bias correction is applied
to the entire timeseries so that bias % in the training period is 0.
}
