% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_region.R
\name{optimise_region}
\alias{optimise_region}
\title{Computes and regionalizes optimal estimate from multiple discharge 
timeseries against observed streamflow}
\usage{
optimise_region(
  HS,
  routing = "instant",
  train = 0.5,
  optim_method = "CLS",
  combination = "ts",
  sampling = "random",
  region_type = "upstream",
  no_station = "em",
  drop = TRUE,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{HS}{An \code{HS} object with observation_ts and runoff_ts}

\item{routing}{Routing algorithm to use. See 
\code{\link{accumulate_runoff}} for options.}

\item{train}{The share of timeseries used for training period.}

\item{optim_method}{A character object giving the optimisation method to be 
used, or a function to be passed to  \code{\link[stats]{optim}}. 
See details.}

\item{combination}{Whether to do the forecast combination for the entire
timeseries, or each month of the year individually, or for full calendar 
years. Accepts \code{"timeseries"}, \code{"ts"}, or \code{"monthly"}, 
\code{"mon"}, or \code{"annual"}, \code{"ann"}.}

\item{sampling}{How to sample training and testing periods. \code{"series"}
for training serially from beginning, or \code{"random"} for a random
sample for both training and testing periods.}

\item{region_type}{How to regionalize combination weights. See details.}

\item{no_station}{How to handle river segments with no downstream 
stations. See details.}

\item{drop}{Drop existing timeseries (columns) in \code{runoff_ts}, 
\code{discharge_ts}, or not.}

\item{...}{Additional parameters passed to the \code{routing} algorithm 
and/or to optimisation..}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns an \code{HS} object with routed and optimised 
\code{discharge_ts}, and additional optimisation information in 
\code{Optimisation_info} and \code{Optimised_at}.
}
\description{
Algorithm routes runoff using user specified routing algorithm, for
each observation point in sequence starting from the most upstream
point. The optimum weight combination is regionalized to the river segments 
upstream from the optimisation point which do not contain observation 
information.
}
\details{
See \code{\link{optimise_point}} for help with optimisation options. Note:
unlike \code{optimise_point}, runoff does not need to be routed in advance,
but it is done by the function automatically.

Currently only one regionalization type is implemented: \code{
upstream} uses the weights obtained for the nearest downstream
station for the river segment.

Options for \code{no_station} is currently only one: \code{"em"}, 
which stands for "ensemble mean". For those river segments with no
weights assigned from a station (segments with no downstream observation 
station), an ensemble mean (each discharge prediction is given equal weight) 
is computed.
}
