% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_HSgrid.R
\name{raster_to_HSgrid}
\alias{raster_to_HSgrid}
\title{Polygonizes a raster and adds cell values as attribute columns to 
an 'sf' polygon object.}
\usage{
raster_to_HSgrid(raster, date, timestep = NULL, aoi = NULL, name = NULL)
}
\arguments{
\item{raster}{A Raster* object.}

\item{timestep}{Length of timestep in raster, e.g. \code{"hour"}, 
\code{"day"}, \code{"month"}}

\item{aoi}{Area of interest as an 'sf' polygon object, 
'SpatialPolygons', or 'SpatialPolygonsDataFrame'. Optional, but 
recommended for all applications.}
}
\value{
Returns an 'sf' polygon object, with each polygon representing 
  a raster cell, and which has been cropped to the area of interest. 
  Columns are ID, area of the polygon in m^2, and each timestep as separate
  column. Output has class 'HSgrid'.
}
\description{
Polygonizes a raster and adds cell values as attribute columns to 
an 'sf' polygon object.
}
