% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_HSgrid.R
\name{raster_to_HSgrid}
\alias{raster_to_HSgrid}
\title{Create HSgrid from a raster}
\usage{
raster_to_HSgrid(rasters, date, timestep = NULL, aoi = NULL,
  names = NULL, verbose = FALSE)
}
\arguments{
\item{rasters}{A raster object, or a list of raster objects, or a list of
URIs to files readable by \code{raster} package.}

\item{date}{Date of the first layer in \code{raster}, or a vector of dates 
with the length of layers in \code{raster}, or a list of vectors of dates,
corresponding to the dates in each raster layer.}

\item{timestep}{Length of timestep in raster. Currently supporting 
\code{"hour"}, \code{"day"}, \code{"month"}. Only needed if \code{date} is
not a vector of dates.}

\item{aoi}{Area of interest as an 'sf' polygon object, 
'SpatialPolygons', or 'SpatialPolygonsDataFrame'. Optional, but 
recommended for all applications.}

\item{names}{A name (names) for the runoff timeserie(s).}

\item{verbose}{Print progress indication or not.}
}
\value{
See \code{\link{create_HSgrid}}.
}
\description{
Polygonizes a raster and adds the values in the raster stack to a data frame
in list column \code{runoff_ts}.
}
