% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_river.R
\name{river_hierarchy}
\alias{river_hierarchy}
\title{Computes different river hierarchies}
\usage{
river_hierarchy(river, ID = "ID", type = "strahler")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network. Required.}

\item{ID}{Name of the column with unique river segment identifiers. Defaults to "ID".}

\item{type}{Type hierarchy to compute. Currently only "strahler" stream order implemented.}
}
\value{
Returns the river network with added column with the selected river hierarchy.
}
\description{
Computes different river hierarchies
}
