% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_hierarchy.R
\name{river_hierarchy}
\alias{river_hierarchy}
\title{Computes river hierarchies for a HSnetwork object.}
\usage{
river_hierarchy(river, type = "strahler", riverID = "riverID")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network.}

\item{type}{Type hierarchy to compute. Currently only \code{strahler} stream order implemented.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to "riverID".}
}
\value{
Returns the river network with added column with the selected river hierarchy.
}
\description{
Computes river hierarchies from a routed river network (\emph{HSnetwork} object), or from an unrouted,
connected river network. In case the input river network is unrouted, \code{\link{river_network}} is run
before computing river hierarchy.
}
\examples{
\dontrun{
library(hydrostreamer)
library(dplyr)

# load data
data(river)

# without prior routing
river <- river_hierarchy(river, riverID="ID")

# with routed network
routed_river <- flow_network(river, riverID = "ID") \%>\%
    river_hierarchy()
}

}
