% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{river_network}
\alias{river_network}
\title{Generates routing information for river routing algorithms}
\usage{
river_network(river, riverID = "riverID", verbose = FALSE)
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to "riverID".}

\item{verbose}{Whether or not print progress indicators.}
}
\value{
Returns the river network with class 'HSnetwork' with the following attributes:
\itemize{
  \item riverID: Unique (renamed) river segment identifier.
  \item PREVIOUS: a list of riverID(s) of the previous river segment(s).
  \item NEXT: riverID of the segment where the river flows to.
  \item UP_SEGMENTS: the number of river segments upstream.
}
}
\description{
The function adds connectivity information to the input \emph{river} object. 
The function assumes that the input river is a "clean" river network: 
connected segments share a node at the start or end of the segment. 
Digitizing direction is required to be the direction of flow.
}
\examples{
\dontrun{
library(hydrostreamer)

# load data
data(river)

routed_river <- river_network(river, riverID = "ID", verbose=TRUE)

}

}
