% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_watershed.R
\name{river_outlets}
\alias{river_outlets}
\title{Identify locations of raster cells which flow in to a river confluence point.}
\usage{
river_outlets(drain.dir, river)
}
\arguments{
\item{drain.dir}{A RasterLayer object of drainage directions.}

\item{river}{An 'sf' linestring feature representing a river network. Required.}
}
\value{
Returns an 'sf' point layer of identified raster cell centers.
}
\description{
The function identifies the coordinates of raster cells which are just upstream from river confluence
points, identified from a river network layer.
}
\section{Details}{

The logic is such that it looks at the last node-pair of each segment in the river network. The angle
of the last node pair is computed, and the raster cell in the direction of the river segment is
identified.

Note that the logic of the function requires that the river network is derived from the same (or at least
the same resolution) raster than the given drainage directions.
}

