% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_network_at_bifurcations.R
\name{split_network_at_bifurcations}
\alias{split_network_at_bifurcations}
\title{Split a HS network into spatially distinct chunks for iterative constant
velocity routing}
\usage{
split_network_at_bifurcations(
  sf_river_network,
  riverID = "riverID",
  verbose = FALSE
)
}
\arguments{
\item{sf_river_network}{A LINESTRING simple feature collection.}

\item{riverID}{unique edge identifier.}

\item{verbose}{Boolean. Print messages.}
}
\value{
Returns the river network with an additional column denoting the
  topological order by which routing must occur.
}
\description{
A method for splitting vector river networks into spatially distinct chunks
for more efficient runoff routing of anabranching networks than simply
running hydrostreamer::accumulate_runoff(routing_method = "constant").
}
\details{
This function reorganizes an HS river object by topology and
  generates topologically ordered groups. The funtion identifies each
  bifurcation point in the network and splits the network at these points
  using a small buffer (10cm). Then the function identifies the resulting
  spatially distinct groups. The groups are then ordered topologically (each
  is assigned a unique id; 1 being furthest upstream and so on), restoring
  the original LINESTRING geometries prior to splitting with buffers.
}
\examples{
# Miera example
library(dplyr)

 miera = hydrostreamer::miera_rivers \%>\%
     dplyr::mutate("default_order" = dplyr::row_number()) \%>\%
     hydrostreamer::river_network(riverID = "riverID",
                                  verbose = TRUE)

 # miera \%>\%
 # ggplot()+
 #    geom_sf(aes(geometry = geom, col = default_order))

 miera_split_at_bifurcations <- 
     split_network_at_bifurcations(miera,
   riverID = "riverID",
   verbose = TRUE)
   
\dontrun{
library(ggplot2)

# individual edges
miera_split_at_bifurcations \%>\%
    ggplot()+
    geom_sf(aes(geometry = geom, col = topo_order)) +
    scale_color_gradientn(colours = c("blue","white","red"))

# groups
miera_split_at_bifurcations \%>\%
    ggplot()+
    geom_sf(aes(geometry = geom, col = river_group_new)) +
    scale_color_gradientn(colours = c("blue","white","red"))
}
}
