% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_river_with_grid.R
\name{split_river_with_grid}
\alias{split_river_with_grid}
\title{Split river segments at runoff unit (polygon) boundaries.}
\usage{
split_river_with_grid(river, HS, riverID = "riverID", zoneID = "zoneID")
}
\arguments{
\item{river}{An 'sf' linestring feature representing a river network.}

\item{HS}{A 'HS' object, obtained with \code{\link{raster_to_HS}}.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to 
\code{"riverID"}.}

\item{zoneID}{Name of the column in \code{HS} with unique IDs.}
}
\value{
Returns an 'sf' linestring object which has been split at 
  the polygon (grid) boundaries with attributes (columns):
  \itemize{
    \item \emph{ID}. Unique ID of the split river segments.
    \item \emph{riverID}. ID of the original river segment prior 
      to splitting.
    \item \emph{zoneID}. ID of the runoff unit split river segment 
      is contained in. 
    \item Other columns inherited from \code{river}.
}
}
\description{
Splits an 'sf' linestring object at the boundaries of 
runoff units (polygons).
}
\examples{
\dontrun{
library(raster)
library(hydrostreamer)

# load data
data(river)
data(basin)
runoff <- brick(system.file("extdata", "runoff.tif", 
                            package = "hydrostreamer"))

# create HS
grid <- raster_to_HS(grid, aoi=basin)

splitriver <- split_river_with_grid(river, grid, 
                                    riverID="ID")
}

}
