% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream.R
\name{upstream}
\alias{upstream}
\title{Find all upstream segments from a specified river segment.}
\usage{
upstream(HSnetwork, ID, riverID = "riverID")
}
\arguments{
\item{HSnetwork}{a \code{HSnetwork} (routed river network) object obtained
with \code{\link{river_network}}.}

\item{ID}{The ID of the segment for which the upstream rivershed is returned.}

\item{riverID}{A character string which specifies the name of the column in 
\code{river} containing unique river network identifiers. Defaults to "riverID".}
}
\value{
A \code{HSnetwork} object (a routed river network) consisting of the 
segments upstream from the segment specified.
}
\description{
Function returns the river network consisting of all upstream segments 
from a specified river segment.
}
